+++ packages/mnemopi/src/index.ts
@@ -7,6 +7,7 @@ export {
 	addMemory,
 	flushExtractions,
 	forget,
+	forgetFact,
 	get,
 	getBank,
 	getContext,